Support both version 1 and 2 synthdef specs
I noticed that trying to read a more complex synthdef wasn't working and
went in search of the a better spec. It turns out there were breaking
changes made to the format between versions 1 and 2. Overtone uses
version 1 by default and that's all I had tested up to now.

This fix means it should work for both versions and contains specs for
both.